pyproject.toml
[tool.poetry]
name = "religious_ecologies"
version = "0.1.0"
description = "Religious Ecologies: Django project for the Religious Ecologies project."
authors = ["Jason Heppler <jheppler@gmu.edu>"]
license = "Mozilla Public License"
readme = "README.md"
#requires-python = ">=3.12"
packages = [{include = "religious_ecologies"}]

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
daphne = "^4.1.2"
whitenoise = "^6.8.2"
django = "^5.1.4"
django-environ = "^0.11.2"
# allauth
django-allauth = {version="^65.3.1", extras=["github"]}

django-tailwind = "^3.8.0"

# postgres
psycopg2 = "^2.9.10"

# obj storage
django-storages = "^1.14.4"
boto3 = "^1.35.94"
django-unfold = "^0.45.0"
django-simple-history = "^3.7.0"
requests = "^2.32.3"
django-dbml = "^0.10.1"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
djhtml = "^3.0.7"
pre-commit = "^4.0.1"
django-debug-toolbar = "^4.4.6"
pytest-django = "^4.9.0"
pylint-django = "^2.6.1"

